[BE] 커스텀 체크리스트 서비스를 분리한다. (#580)
