+++ backend/bang-ggood/src/main/java/com/bang_ggood/question/dto/response/CategoryCustomChecklistQuestionsResponse.java
@@ -3,4 +3,8 @@ package com.bang_ggood.question.dto.response;
 import java.util.List;
 
 public record CategoryCustomChecklistQuestionsResponse(List<CategoryCustomChecklistQuestionResponse> categories) {
+
+    public static CategoryCustomChecklistQuestionsResponse from(List<CategoryCustomChecklistQuestionResponse> categories) {
+        return new CategoryCustomChecklistQuestionsResponse(categories);
+    }
 }
