--- backend/bang-ggood/src/main/java/com/bang_ggood/question/dto/response/ChecklistQuestionsResponse.java
@@ -1,6 +1,0 @@
-package com.bang_ggood.question.dto.response;
-
-import java.util.List;
-
-public record ChecklistQuestionsResponse(List<CategoryQuestionsResponse> categories) {
-}
