+++ backend/bang-ggood/src/main/java/com/bang_ggood/question/service/ChecklistQuestionService.java
@@ -3,7 +3,11 @@ package com.bang_ggood.question.service;
 import com.bang_ggood.global.exception.BangggoodException;
 import com.bang_ggood.global.exception.ExceptionCode;
 import com.bang_ggood.question.domain.ChecklistQuestion;
+import com.bang_ggood.question.domain.CustomChecklistQuestion;
+import com.bang_ggood.question.domain.Question;
 import com.bang_ggood.question.repository.ChecklistQuestionRepository;
+import com.bang_ggood.question.repository.CustomChecklistQuestionRepository;
+import com.bang_ggood.user.domain.User;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import java.util.HashSet;
@@ -14,9 +18,11 @@ import java.util.Set;
 public class ChecklistQuestionService {
 
     private final ChecklistQuestionRepository checklistQuestionRepository;
+    private final CustomChecklistQuestionRepository customChecklistQuestionRepository;
 
-    public ChecklistQuestionService(ChecklistQuestionRepository checklistQuestionRepository) {
+    public ChecklistQuestionService(ChecklistQuestionRepository checklistQuestionRepository, CustomChecklistQuestionRepository customChecklistQuestionRepository) {
         this.checklistQuestionRepository = checklistQuestionRepository;
+        this.customChecklistQuestionRepository = customChecklistQuestionRepository;
     }
 
     @Transactional
@@ -32,5 +38,35 @@ public class ChecklistQuestionService {
                 throw new BangggoodException(ExceptionCode.QUESTION_DUPLICATED);
             }
         });
+    }
+
+    @Transactional(readOnly = true)
+    public List<CustomChecklistQuestion> readCustomChecklistQuestions(User user) {
+        return customChecklistQuestionRepository.findAllByUser(user);
+    }
+
+    @Transactional
+    public void updateCustomChecklist(User user, List<Question> questions) {
+        validateCustomChecklistQuestionsIsNotEmpty(questions);
+        validateCustomChecklistQuestionsDuplication(questions);
+
+        customChecklistQuestionRepository.deleteAllByUser(user);
+
+        List<CustomChecklistQuestion> customChecklistQuestions = questions.stream()
+                .map(question -> new CustomChecklistQuestion(user, question))
+                .toList();
+        customChecklistQuestionRepository.saveAll(customChecklistQuestions);
+    }
+
+    private void validateCustomChecklistQuestionsIsNotEmpty(List<Question> questions) {
+        if (questions.isEmpty()) {
+            throw new BangggoodException(ExceptionCode.CUSTOM_CHECKLIST_QUESTION_EMPTY);
+        }
+    }
+
+    private void validateCustomChecklistQuestionsDuplication(List<Question> questions) {
+        if (questions.size() != Set.copyOf(questions).size()) {
+            throw new BangggoodException(ExceptionCode.QUESTION_DUPLICATED);
+        }
     }
 }
