+++ backend/bang-ggood/src/test/java/com/bang_ggood/auth/service/AuthServiceTest.java
@@ -8,7 +8,8 @@ import com.bang_ggood.global.exception.BangggoodException;
 import com.bang_ggood.global.exception.ExceptionCode;
 import com.bang_ggood.question.domain.Question;
 import com.bang_ggood.question.dto.response.CategoryQuestionsResponse;
-import com.bang_ggood.question.dto.response.ChecklistQuestionsResponse;
+import com.bang_ggood.question.dto.response.CustomChecklistQuestionsResponse;
+import com.bang_ggood.question.service.QuestionManageService;
 import com.bang_ggood.user.UserFixture;
 import com.bang_ggood.user.domain.User;
 import com.bang_ggood.user.repository.UserRepository;
@@ -37,6 +38,8 @@ class AuthServiceTest extends IntegrationTestSupport {
     private AuthService authService;
     @Autowired
     private ChecklistService checklistService;
+    @Autowired
+    private QuestionManageService questionManageService;
     @Autowired
     private UserRepository userRepository;
     @Autowired
@@ -83,10 +86,10 @@ class AuthServiceTest extends IntegrationTestSupport {
 
         // then
         User user = authService.extractUser(token);
-        ChecklistQuestionsResponse checklistQuestions = checklistService.readChecklistQuestions(user);
+        CustomChecklistQuestionsResponse customChecklistQuestions = questionManageService.readCustomChecklistQuestions(user);
 
         int sum = 0;
-        for (CategoryQuestionsResponse response : checklistQuestions.categories()) {
+        for (CategoryQuestionsResponse response : customChecklistQuestions.categories()) {
             sum += response.questions().size();
         }
